src_for_windows/requirements.txt
vincenty==0.1.4
opencv_python==3.4.14.53
utm==0.7.0
geopy==2.1.0
scipy==1.4.0
Shapely==1.7.0
descartes==1.1.0
pyexifinfo==0.4.0
imutils==0.5.3
scikit_image==0.17.2
matplotlib==3.1.2
numpy==1.17.3
Pillow==8.2.0
weathercom
requests
